requirements.txt
cog==0.9.9
detail==0.2.4
fvcore==0.1.5.post20221221
iopath==0.1.9
matplotlib==3.8.1
numpy==1.26.4
open_clip_torch==2.20.0
opencv_python==4.8.1.78
Pillow==10.3.0
pycocotools==2.0.7
scipy==1.13.1
setuptools==60.2.0
Shapely==2.0.4
tabulate==0.9.0
termcolor==2.4.0
timm==0.9.8
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.2
wandb==0.15.10
